Refactor GUI handling and timer management; migrate GUI creation to components and implement game pause/resume functionality
